… app/(user)/contribute/_components/qr-extraction-feature.tsx
… app/(user)/contribute/_components/qr-processor.tsx
… app/(user)/leaderboard/_components/async-top-contributors.tsx
… app/(user)/leaderboard/_components/lazy-leaderboard-content.tsx
… app/(user)/leaderboard/_components/leaderboard-content.tsx
… app/(user)/leaderboard/_lib/queries.ts
… app/(user)/leaderboard/page.tsx
… app/(user)/my-contributions/_components/async-contribution-list.tsx
… app/(user)/my-contributions/_components/async-stats-cards.tsx
… app/(user)/my-contributions/_components/loading-skeletons.tsx
… app/(user)/my-contributions/_lib/queries.ts
… app/(user)/my-contributions/page.tsx
… docs/user-features-optimization/claude.md
… docs/user-features-optimization/findings.md
… docs/user-features-optimization/gemini.md
… docs/user-features-optimization/plan.md
… hooks/use-location-prefill-lazy.ts
… hooks/use-qr-extraction-lazy.ts